Control Functions added
In this commit, what I have done is added functions that will control an object(the robot in our case). Me and Jash hae not created the robot yet, but we have created the functions to control it through the xbox controller. The functions include every type of movement: left, right, up, down , and all the directions diagonally as well.