Cleaning Up Code, Fixing Method Structure, and Commenting
I just commented code to make it look better. Also, I added a message in the robot file letting everyone know our status. The movement code is done other than we have to implement it with robot controls (going to go int the robotMovement method).  We can not start on Intake or Shooting methods just yet until we get more done on the environment/physics end. Lastly, for the robot design itself me and Arash will figure that out later on (right now it is just a purple square).